+++ app/public/scripts/yummy/cms/session/deleteSession.js
@@ -1,4 +1,10 @@
 async function handleDeleteSession(sessionId) {
+
+    const reservationsData = await loadReservationRelatedData(sessionId);
+    if (reservationsData !== null) {
+        const userConfirmation = confirm(`WARNING! Deleting this session will also deactivate ${reservationsData.numberOfReservations} reservations for ${reservationsData.numberOfPeople} people in ${reservationsData.numberOfRestaurants} restaurants.`);
+    if (!userConfirmation) return;
+    }
     const response = await fetch(`/api/session?id=${sessionId}`, {
         method: "DELETE",
         headers: {
@@ -11,4 +17,11 @@ async function handleDeleteSession(sessionId) {
 
     itemsListContainer.innerHTML = "";
     loadItems(sessionAPIendpoint, "session")
+}
+
+async function loadReservationRelatedData(sessionId){
+    const reservations = await fetch(`${reservationAPIendpoint}?sessionId=${sessionId}`);
+    const reservationsData = await reservations.json();
+
+    return reservationsData;
 }
